+++ mis_builder/models/mis_report_style.py
@@ -74,16 +74,6 @@ class MisReportKpiStyle(models.Model):
         ("xx-large", "xx-large"),
     ]
 
-    _font_size_to_xlsx_size = {
-        "medium": 11,
-        "xx-small": 5,
-        "x-small": 7,
-        "small": 9,
-        "large": 13,
-        "x-large": 15,
-        "xx-large": 17,
-    }
-
     # style name
     name = fields.Char(string="Style name", required=True)
 
@@ -308,36 +298,6 @@ class MisReportKpiStyle(models.Model):
         if delta is not AccountingNone:
             delta_r = self.render(lang, delta_style, delta_type, delta, sign="+")
         return delta, delta_r, delta_style, delta_type
-
-    @api.model
-    def to_xlsx_style(self, var_type, props, no_indent=False):
-        xlsx_attributes = [
-            ("italic", props.font_style == "italic"),
-            ("bold", props.font_weight == "bold"),
-            ("font_size", self._font_size_to_xlsx_size.get(props.font_size, 11)),
-            ("font_color", props.color),
-            ("bg_color", props.background_color),
-        ]
-        if var_type == TYPE_NUM:
-            num_format = "#,##0"
-            if props.dp:
-                num_format += "."
-                num_format += "0" * props.dp
-            if props.prefix:
-                num_format = f'"{props.prefix} "{num_format}'
-            if props.suffix:
-                num_format = f'{num_format}" {props.suffix}"'
-            xlsx_attributes.append(("num_format", num_format))
-        elif var_type == TYPE_PCT:
-            num_format = "0"
-            if props.dp:
-                num_format += "."
-                num_format += "0" * props.dp
-            num_format += "%"
-            xlsx_attributes.append(("num_format", num_format))
-        if props.indent_level is not None and not no_indent:
-            xlsx_attributes.append(("indent", props.indent_level))
-        return dict([a for a in xlsx_attributes if a[1] is not None])
 
     @api.model
     def to_css_style(self, props, no_indent=False):
