+++ mis_builder/report/__init__.py
@@ -2,4 +2,3 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
 from . import mis_report_instance_qweb
-from . import mis_report_instance_xlsx
