+++ mis_builder/views/mis_report_instance.xml
@@ -35,12 +35,6 @@
                     icon="fa-search"
                 />
                 <button type="object" name="print_pdf" string="Print" icon="fa-print" />
-                <button
-                    type="object"
-                    name="export_xls"
-                    string="Export"
-                    icon="fa-download"
-                />
                 <field name="name" />
                 <field name="report_id" string="Template" />
                 <field name="company_id" groups="base.group_multi_company" />
@@ -72,13 +66,6 @@
                             name="print_pdf"
                             string="Print"
                             icon="fa-print"
-                            class="btn"
-                        />
-                        <button
-                            type="object"
-                            name="export_xls"
-                            string="Export"
-                            icon="fa-download"
                             class="btn"
                         />
                         <button
@@ -271,13 +258,6 @@
                     name="print_pdf"
                     string="Print"
                     icon="fa-print"
-                    class="btn"
-                />
-                    <button
-                    type="object"
-                    name="export_xls"
-                    string="Export"
-                    icon="fa-download"
                     class="btn"
                 /> or <button string="Cancel" class="oe_link" special="cancel" />
                 </footer>
