+++ mis_builder_report_xlsx/README.rst
@@ -1,0 +1,112 @@
+=========================
+MIS Builder - Report XLSX
+=========================
+
+.. 
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! source digest: sha256:5c01f05913a972ff6c782fea6e1fe238222c3296f4d7ae10e4f2f86b2cf11fbe
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fmis--builder-lightgray.png?logo=github
+    :target: https://github.com/OCA/mis-builder/tree/18.0/mis_builder_report_xlsx
+    :alt: OCA/mis-builder
+.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
+    :target: https://translation.odoo-community.org/projects/mis-builder-18-0/mis-builder-18-0-mis_builder_report_xlsx
+    :alt: Translate me on Weblate
+.. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
+    :target: https://runboat.odoo-community.org/builds?repo=OCA/mis-builder&target_branch=18.0
+    :alt: Try me on Runboat
+
+|badge1| |badge2| |badge3| |badge4| |badge5|
+
+This module improves the MIS Reports modules, allowing to downloads MIS
+Reports in XLSX format.
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Usage
+=====
+
+- Go to MIS Reports list and form views.
+- Click on 'Export' Button to download your XLSX file.
+
+|image|
+
+- In the same way, a button is present in the 'preview' form view of
+  your MIS reports.
+
+|image1|
+
+.. |image| image:: https://raw.githubusercontent.com/OCA/mis-builder/18.0/mis_builder_report_xlsx/static/description/mis_report_instance_list.png
+.. |image1| image:: https://raw.githubusercontent.com/OCA/mis-builder/18.0/mis_builder_report_xlsx/static/description/mis_report_instance_preview_form.png
+
+Known issues / Roadmap
+======================
+
+The mis_builder
+`roadmap <https://github.com/OCA/mis-builder/issues?q=is%3Aopen+is%3Aissue+label%3Aenhancement>`__
+and `known
+issues <https://github.com/OCA/mis-builder/issues?q=is%3Aopen+is%3Aissue+label%3Abug>`__
+can be found on github.
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/mis-builder/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us to smash it by providing a detailed and welcomed
+`feedback <https://github.com/OCA/mis-builder/issues/new?body=module:%20mis_builder_report_xlsx%0Aversion:%2018.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+-------
+
+* ACSONE SA/NV
+* GRAP
+
+Contributors
+------------
+
+- Stéphane Bidoul <stephane.bidoul@acsone.eu>
+- Sylvain LE GAL
+
+Maintainers
+-----------
+
+This module is maintained by the OCA.
+
+.. image:: https://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: https://odoo-community.org
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
+
+.. |maintainer-sbidoul| image:: https://github.com/sbidoul.png?size=40px
+    :target: https://github.com/sbidoul
+    :alt: sbidoul
+
+Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
+
+|maintainer-sbidoul| 
+
+This module is part of the `OCA/mis-builder <https://github.com/OCA/mis-builder/tree/18.0/mis_builder_report_xlsx>`_ project on GitHub.
+
+You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
