+++ mis_builder_report_xlsx/readme/CONTRIBUTORS.md
@@ -1,0 +1,2 @@
+- Stéphane Bidoul \<<stephane.bidoul@acsone.eu>\>
+- Sylvain LE GAL
