+++ mis_builder_report_xlsx/readme/DESCRIPTION.md
@@ -1,0 +1,2 @@
+This module improves the MIS Reports modules, allowing to
+downloads MIS Reports in XLSX format.
