+++ mis_builder_report_xlsx/readme/USAGE.md
@@ -1,0 +1,10 @@
+- Go to MIS Reports list and form views.
+- Click on 'Export' Button to download your XLSX file.
+
+![image](../static/description/mis_report_instance_list.png)
+
+
+- In the same way, a button is present in the 'preview' form view
+  of your MIS reports.
+
+![image](../static/description/mis_report_instance_preview_form.png)
