+++ mis_builder_report_xlsx/report/__init__.py
@@ -1,0 +1,1 @@
+from . import mis_report_instance_xlsx
