+++ mis_builder_report_xlsx/tests/__init__.py
@@ -1,0 +1,2 @@
+from . import test_mis_report_instance
+from . import test_render
